move trickLogic tests into correct folder
